Move fake_pool_server fixture from test file to conftest.py
Problem: The fake_pool_server pytest fixture was defined inline in
host_pool_client_test.py, violating the project convention that all
fixtures must live in conftest.py.

Fix: Moved the _FAKE_LEASE_RESPONSE constant, _FakePoolHandler class,
and fake_pool_server fixture to conftest.py, and removed the
corresponding definitions and now-unused imports from the test file.
